バックエンドのリファクタリング
