+++ .env.example
@@ -1,1 +1,2 @@
-DATABASE_URL=
+POSTGRESQL_DATABASE_URL=
+OPENAI_API_KEY=
