+++ app/dependencies.py
@@ -1,6 +1,4 @@
-from sqlalchemy.orm import Session
-from app.database import SessionLocal
-from fastapi import Depends
+from app.database.database import SessionLocal
 
 
 def get_db():
