+++ app/models.py
@@ -1,6 +1,5 @@
-from pydantic import BaseModel
 from sqlalchemy import Column, Integer, String
-from app.database import Base
+from app.database.database import Base
 
 
 class Job(Base):
